Implemented basic recording functionality
Added the recording functionality to the dashcam fragment kotlin and
xml.
Added drawables for use in the dashcam fragment ui.
Added colors for use in the dashcam fragment ui.
Added strings for use in the dashcam fragment ui.

Added required dependencies to the gradle app file.

It will save recorded videos with a timestamp in the filename to prevent
overwriting old files. The saved video files can be accessed with ADB in
`/data/data/com.camerasecuritysystem.client/files`.