feat: Add settings input validation
Notable changes:
- The TextInputLayouts now have errors enabled. The password layout
was also missing an ID. The camera and password layouts get a proper
hint. The IP address layout is now a text input instead of phone
number input. The Camera ID input is a number input instead of a text.
- When the alert dialog is first created, the values from the Shared
Preferences are loaded and checked if they are null. If they are, the
appropriate error message will display and the OK button will be
disabled. This button is only enabled if all fields are validated.
- When text in the input fields changes, the new value is checked for
validation. If they are valid, the error is cleared.
- When applying the text from the alert dialog, the values are not
checked anymore, since the OK button can only be pressed when the
values are valid and thus non-null.