feat/refac: Reuse client to allow disconnecting
Notable changes:
- Globalscope.launch has been moved to the ServerConnection class to
simplify its usage.
- Don't call client.close(), otherwise the client will have to be
recreated. Closing the websocketChannel is enough.
- Allow disconnecting when succesfully connected. The button also
reflects the connecting stage. When no credentials are entered, the
button is invisible.
- Two utility functions have been implemented to facilitate debug mode
assertions (only enabled in debug mode). These are used to verify the
websocket (dis)connecting procedure.
- To communicate the state of the connection to activities, a LiveData
implementation is used to allow for callbacks to be executed in the
relevant activity.