fix: Don't fail release build on example leftovers
When building in release mode, the build will fail with a linting
error: MissingDefaultResource. This is caused by there being no
dimens.xml file in the values folder, while there were dimens.xml
files in the values-land, etc. folders. This is a leftover from the
example project, which actually used FloatingActionButton margins.